+++ ToyKeeper/crescendo/misc.h
@@ -8,7 +8,7 @@ void _delay_input();
 void blink(uint8_t val, uint8_t on_4ms, uint8_t off_4ms);
 #define buzz() blink(32, 500 / 32 / 4, 500 / 32 / 2);
 #ifdef USE_BLINK_NUM
-void blink_num(uint8_t num);
+void blink_num(uint16_t num);
 #endif
 void next_mode();
